+++ execution/destroy.sh
