+++ execution/src/main/java/executor/service/exception/ScenarioExecutionException.java
@@ -1,0 +1,10 @@
+package executor.service.exception;
+
+public class ScenarioExecutionException extends RuntimeException {
+    public ScenarioExecutionException(String message, Throwable cause) {
+        super(message, cause);
+    }
+    public ScenarioExecutionException(String message) {
+        super(message);
+    }
+}
