+++ execution/src/main/java/executor/service/exception/step/ClickCssException.java
@@ -1,0 +1,8 @@
+package executor.service.exception.step;
+
+
+public class ClickCssException extends StepExecutionException {
+    public ClickCssException(Throwable cause) {
+        super("Failed to perform 'clickCss' step.", cause);
+    }
+}
