+++ execution/src/main/java/executor/service/exception/step/ClickXPathException.java
@@ -1,0 +1,8 @@
+package executor.service.exception.step;
+
+
+public class ClickXPathException extends StepExecutionException {
+    public ClickXPathException(Throwable cause) {
+        super("Failed to perform 'clickXpath' step.", cause);
+    }
+}
