+++ execution/src/main/java/executor/service/exception/step/SleepException.java
@@ -1,0 +1,8 @@
+package executor.service.exception.step;
+
+
+public class SleepException extends StepExecutionException {
+    public SleepException(Throwable cause) {
+        super("Failed to perform 'Sleep' step.", cause);
+    }
+}
