+++ execution/src/main/java/executor/service/exception/step/StepExecutionException.java
@@ -1,0 +1,12 @@
+package executor.service.exception.step;
+
+import executor.service.exception.ScenarioExecutionException;
+
+public class StepExecutionException extends ScenarioExecutionException {
+    public StepExecutionException(String message, Throwable cause) {
+        super(message, cause);
+    }
+    public StepExecutionException(String message) {
+        super(message);
+    }
+}
