+++ execution/src/main/java/executor/service/exception/step/UnknownStepException.java
@@ -1,0 +1,7 @@
+package executor.service.exception.step;
+
+public class UnknownStepException extends StepExecutionException {
+    public UnknownStepException(String message) {
+        super(message);
+    }
+}
