--- executor-service/src/main/java/executor/service/exception/scenario/step/ClickCssException.java
@@ -1,8 +1,0 @@
-package executor.service.exception.scenario.step;
-
-
-public class ClickCssException extends StepExecutionException {
-    public ClickCssException(Throwable cause) {
-        super("Failed to perform 'clickCss' step.", cause);
-    }
-}
