--- executor-service/src/main/java/executor/service/exception/scenario/step/ClickXPathException.java
@@ -1,8 +1,0 @@
-package executor.service.exception.scenario.step;
-
-
-public class ClickXPathException extends StepExecutionException {
-    public ClickXPathException(Throwable cause) {
-        super("Failed to perform 'clickXpath' step.", cause);
-    }
-}
