--- executor-service/src/main/java/executor/service/exception/scenario/step/SleepException.java
@@ -1,8 +1,0 @@
-package executor.service.exception.scenario.step;
-
-
-public class SleepException extends StepExecutionException {
-    public SleepException(Throwable cause) {
-        super("Failed to perform 'Sleep' step.", cause);
-    }
-}
