--- executor-service/src/main/java/executor/service/exception/scenario/step/UnknownStepException.java
@@ -1,7 +1,0 @@
-package executor.service.exception.scenario.step;
-
-public class UnknownStepException extends StepExecutionException {
-    public UnknownStepException(String message) {
-        super(message);
-    }
-}
