--- executor-service/src/main/java/executor/service/execution/scenario/step/Sleep.java
@@ -1,37 +1,0 @@
-package executor.service.execution.scenario.step;
-
-import executor.service.logger.annotation.Logged;
-import executor.service.exception.scenario.step.SleepException;
-import executor.service.model.Step;
-import org.openqa.selenium.WebDriver;
-import org.springframework.stereotype.Component;
-
-import java.util.Random;
-@Component
-@Logged
-public class Sleep implements StepExecution {
-    @Override
-    public String getStepAction() {
-        return "sleep";
-    }
-
-    @Override
-    public void step(WebDriver webDriver, Step step) {
-        try {
-            Thread.sleep(getDuration(step));
-        } catch (InterruptedException e) {
-            throw new SleepException(e);
-        }
-    }
-
-    private int getDuration(Step step) {
-        try {
-            String[] values = step.getValue().split(":");
-            int first = Integer.parseInt(values[0]);
-            int second = Integer.parseInt(values[1]);
-            return new Random().nextInt(second - first + 1) + first;
-        } catch (ArrayIndexOutOfBoundsException | IllegalArgumentException ex) {
-            throw new SleepException(ex);
-        }
-    }
-}
